% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{list_clonal_mu}
\alias{list_clonal_mu}
\title{Clonal Mutations with Various Parameters and Mu Fixed}
\usage{
list_clonal_mu(n_values, R0_values, mu, maxMuGen, maxFS, clonal)
}
\arguments{
\item{n_values}{A sequence of initial population sizes}

\item{R0_values}{A sequence of reproduction number}

\item{mu}{Mutation rate, a positive number}

\item{maxMuGen}{Maximum number of mutant lineages being calculated}

\item{maxFS}{Maximum number of final sizes being calculated, equal to or smaller than maxMuGen}

\item{clonal}{Maximum number of clonal mutations being calculated}
}
\value{
A list of clonal mutation pmf with different parameters
}
\description{
This function calculates for the clonal mutation distribution
across pairs of parameters N and R0.
}
\examples{
list_clonal_mu(1:10, 2:5, 0.3, 50, 50, 5)
list_clonal_mu(seq(2, 10, by = 2), seq(2, 4, by = 0.3), 2.1, 60, 50, 7)
}
