% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{LL_lambda_nb}
\alias{LL_lambda_nb}
\title{Lambda Log Likelihood with Negative Binomial as Initial Distribution}
\usage{
LL_lambda_nb(df, listClonal, lambda, R0, mu, maxMuGen, maxFS, maxIni, k)
}
\arguments{
\item{df}{Data set with first column named "ClonalMu" recording number of clonal mutations, and the second column named "freq" recording number of times the clonal mutation is observed}

\item{listClonal}{Returned list from function list_clonal with same parameters}

\item{lambda}{Mean of Poisson distribution, a positive number}

\item{R0}{Reproduction number, a positive number}

\item{mu}{Mutation rate, a positive number}

\item{maxMuGen}{Maximum number of mutant lineages being calculated}

\item{maxFS}{Maximum number of final sizes being calculated, equal to or smaller than maxMuGen}

\item{maxIni}{Maximum initial population size being calculated}

\item{k}{dispersion paramter of negative binomial distribution}
}
\value{
A number representing the probability
}
\description{
This function calculates the log likelihood of getting the observed data
points in the provided data set given a certain lambda, R0, and mu.
}
\examples{
LL_lambda_nb(sample_data, listClonal, 2.1, 1.6, 0.4, 50, 50, 10, 0.1)
LL_lambda_nb(flu_data, listClonal, 0.2, 10, 0.7, 60, 50, 8, 0.5)
}
