% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{LL_prop.df}
\alias{LL_prop.df}
\title{Log Likelihood Heat Map for Proportion}
\usage{
LL_prop.df(df, listClonal, prop_values, R0, mu_values, maxMuGen, maxFS, maxIni)
}
\arguments{
\item{df}{Data set with first column named "ClonalMu" recording number of clonal mutations, and the second column named "freq" recording number of times the clonal mutation is observed}

\item{listClonal}{Returned list from function list_clonal with same parameters}

\item{prop_values}{A sequence of proportion of N=1, between 0 and 1}

\item{R0}{Reproduction number, a positive number}

\item{mu_values}{A sequence of mutation rate}

\item{maxMuGen}{Maximum number of mutant lineages being calculated}

\item{maxFS}{Maximum number of final sizes being calculated, equal to or smaller than maxMuGen}

\item{maxIni}{Maximum initial population size being calculated}
}
\value{
A data frame with proportion, mean N, mu, and corresponding probabilities of getting the set of parameters given data points in the input data set
}
\description{
This function calculates the log likelihood of getting the observed data
points in the provided data set across a set of parameters proportion of N=1,
mean N, and mu.
}
\examples{
LL_prop.df(sample_data, listClonal, seq(0.1, 1, by = 0.2), 1.6, seq(0.1, 2, by = 0.2), 50, 50, 10)
LL_prop.df(flu_data, listClonal, seq(0.1, 0.8, by = 0.1), 10, seq(1.2, 2.5, by = 0.1), 60, 50, 8)
}
