% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{pmf.clonal_params_mu}
\alias{pmf.clonal_params_mu}
\title{Clonal Mutation Probability with Various Parameters and Mu Fixed}
\usage{
pmf.clonal_params_mu(
  clonal,
  listClonal,
  n_values,
  R0_values,
  mu,
  maxMuGen,
  maxFS
)
}
\arguments{
\item{clonal}{The number of clonal mutations being calculated, a positive integer}

\item{listClonal}{Returned list from function list_clonal_mu with same parameters}

\item{n_values}{A sequence of initial population sizes}

\item{R0_values}{A sequence of reproduction number}

\item{mu}{Mutation rate, a positive number}

\item{maxMuGen}{Maximum number of mutant lineages being calculated}

\item{maxFS}{Maximum number of final sizes being calculated, equal to or smaller than maxMuGen}
}
\value{
A data frame of N, R0 values and corresponding probability of getting the input clonal mutations
}
\description{
This function calculates the probability of getting a certain
number of clonal mutations across pairs of parameters N and R0.
}
\examples{
pmf.clonal_params_mu(0, listClonal, 1:10, 2:5, 0.3, 50, 50)
pmf.clonal_params_mu(2, listClonal, seq(2, 10, by = 2), seq(2, 4, by = 0.3), 1.2, 60, 50)
}
