% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{pmf.ini_size_nb}
\alias{pmf.ini_size_nb}
\title{Initial Population Size Distribution as Adjusted Negative Binomial}
\usage{
pmf.ini_size_nb(lambda, R0, maxIni, k)
}
\arguments{
\item{lambda}{Mean of Poisson distribution, a positive number}

\item{R0}{Reproduction number, a positive number}

\item{maxIni}{Maximum initial population size being calculated}

\item{k}{dispersion paramter of negative binomial distribution}
}
\value{
A data frame of N, distribution type and corresponding probability of getting that number of initial population size
}
\description{
This function calculates the negative binomial distribution of initial
population size N given a mean of lambda and also the adjusted distribution
normalized by at least one of the initial viral particles established.
}
\examples{
pmf.ini_size_nb(2.1, 1.6, 10, 0.1)
pmf.ini_size_nb(0.8, 2.4, 6, 0.5)
}
